+++ src/emulator/storage/apis/gcloud.ts
@@ -123,20 +123,14 @@ export function createCloudEndpoints(emulator: StorageEmulator): Router {
       });
     });
 
-    let upload = storageLayer.uploadBytes(uploadId, req.body);
-
+    const upload = storageLayer.uploadBytes(uploadId, req.body);
     if (!upload) {
       res.sendStatus(400);
       return;
     }
 
-    const finalizedUpload = storageLayer.finalizeUpload(uploadId);
-    if (!finalizedUpload) {
-      res.sendStatus(400);
-      return;
-    }
-    upload = finalizedUpload.upload;
-    res.status(200).json(new CloudStorageObjectMetadata(finalizedUpload.file.metadata)).send();
+    const uploadedFile = storageLayer.finalizeUpload(upload);
+    res.status(200).json(new CloudStorageObjectMetadata(uploadedFile.metadata)).send();
   });
 
   gcloudStorageAPI.post("/b/:bucketId/o/:objectId/acl", (req, res) => {