+++ src/emulator/storage/files.ts
@@ -53,6 +53,7 @@ export class ResumableUpload {
   private _bucketId: string;
   private _objectId: string;
   private _contentType: string;
+  private _authorization: string | undefined;
   private _currentBytesUploaded = 0;
   private _status: UploadStatus = UploadStatus.ACTIVE;
   private _fileLocation: string;
@@ -62,13 +63,15 @@ export class ResumableUpload {
     objectId: string,
     uploadId: string,
     contentType: string,
-    metadata: IncomingMetadata
+    metadata: IncomingMetadata,
+    authorization?: string
   ) {
     this._bucketId = bucketId;
     this._objectId = objectId;
     this._uploadId = uploadId;
     this._contentType = contentType;
     this._metadata = metadata;
+    this._authorization = authorization;
     this._fileLocation = encodeURIComponent(`${uploadId}_b_${bucketId}_o_${objectId}`);
     this._currentBytesUploaded = 0;
   }
@@ -90,6 +93,9 @@ export class ResumableUpload {
   }
   public set contentType(contentType: string) {
     this._contentType = contentType;
+  }
+  public get authorization(): string | undefined {
+    return this._authorization;
   }
   public get currentBytesUploaded(): number {
     return this._currentBytesUploaded;
@@ -113,11 +119,6 @@ export enum UploadStatus {
   CANCELLED,
   FINISHED,
 }
-
-export type FinalizedUpload = {
-  upload: ResumableUpload;
-  file: StoredFile;
-};
 
 export class StorageLayer {
   private _files!: Map<string, StoredFile>;
@@ -167,6 +168,27 @@ export class StorageLayer {
     return;
   }
 
+  /**
+   * Generates metadata for an uploaded file. Generally, this should only be used for finalized
+   * uploads, unless needed for security rule checks.
+   * @param upload The upload corresponding to the file for which to generate metadata.
+   * @returns Metadata for uploaded file.
+   */
+  public createMetadata(upload: ResumableUpload): StoredFileMetadata {
+    const bytes = this._persistence.readBytes(upload.fileLocation, upload.currentBytesUploaded);
+    return new StoredFileMetadata(
+      {
+        name: upload.objectId,
+        bucket: upload.bucketId,
+        contentType: "",
+        contentEncoding: upload.metadata.contentEncoding,
+        customMetadata: upload.metadata.metadata,
+      },
+      this._cloudFunctions,
+      bytes
+    );
+  }
+
   public getBytes(
     bucket: string,
     object: string,
@@ -190,10 +212,18 @@ export class StorageLayer {
     bucket: string,
     object: string,
     contentType: string,
-    metadata: IncomingMetadata
+    metadata: IncomingMetadata,
+    authorization?: string
   ): ResumableUpload {
     const uploadId = v4();
-    const upload = new ResumableUpload(bucket, object, uploadId, contentType, metadata);
+    const upload = new ResumableUpload(
+      bucket,
+      object,
+      uploadId,
+      contentType,
+      metadata,
+      authorization
+    );
     this._uploads.set(uploadId, upload);
     return upload;
   }
@@ -202,13 +232,18 @@ export class StorageLayer {
     return this._uploads.get(uploadId);
   }
 
-  public cancelUpload(uploadId: string): ResumableUpload | undefined {
-    const upload = this._uploads.get(uploadId);
-    if (!upload) {
-      return undefined;
+  /**
+   * Deletes partially uploaded file from persistence layer and updates its status. Cancelling
+   * an upload is idempotent.
+   * @param upload The upload to be cancelled.
+   * @returns Whether the upload was cancelled (i.e. its initial status was ACTIVE or CANCELLED).
+   */
+  public cancelUpload(upload: ResumableUpload): boolean {
+    if (upload.status === UploadStatus.ACTIVE) {
+      this._persistence.deleteFile(upload.fileLocation);
+      upload.status = UploadStatus.CANCELLED;
     }
-    upload.status = UploadStatus.CANCELLED;
-    this._persistence.deleteFile(upload.fileLocation);
+    return upload.status === UploadStatus.CANCELLED;
   }
 
   public uploadBytes(uploadId: string, bytes: Buffer): ResumableUpload | undefined {
@@ -252,37 +287,26 @@ export class StorageLayer {
     return this._persistence.deleteAll();
   }
 
-  public finalizeUpload(uploadId: string): FinalizedUpload | undefined {
-    const upload = this._uploads.get(uploadId);
-
-    if (!upload) {
-      return undefined;
-    }
-
+  /**
+   * Stores the uploaded file with generated metadata and triggers Object Finalize Cloud Functions.
+   * @param upload The upload to finalize.
+   * @returns The stored file.
+   */
+  public finalizeUpload(upload: ResumableUpload): StoredFile {
     upload.status = UploadStatus.FINISHED;
-    const filePath = this.path(upload.bucketId, upload.objectId);
 
-    const bytes = this._persistence.readBytes(upload.fileLocation, upload.currentBytesUploaded);
-    const finalMetadata = new StoredFileMetadata(
-      {
-        name: upload.objectId,
-        bucket: upload.bucketId,
-        contentType: "",
-        contentEncoding: upload.metadata.contentEncoding,
-        customMetadata: upload.metadata.metadata,
-      },
-      this._cloudFunctions,
-      bytes,
-      upload.metadata
-    );
-    const file = new StoredFile(finalMetadata, filePath);
+    const metadata = this.createMetadata(upload);
+    const filePath = this.path(upload.bucketId, upload.objectId);
+    const file = new StoredFile(metadata, filePath);
+
     this._files.set(filePath, file);
 
-    this._persistence.deleteFile(filePath, true);
+    this._persistence.deleteFile(filePath, /* failSilently = */ true);
     this._persistence.renameFile(upload.fileLocation, filePath);
 
     this._cloudFunctions.dispatch("finalize", new CloudStorageObjectMetadata(file.metadata));
-    return { upload: upload, file: file };
+
+    return file;
   }
 
   public oneShotUpload(
@@ -306,8 +330,7 @@ export class StorageLayer {
         customMetadata: incomingMetadata.metadata,
       },
       this._cloudFunctions,
-      bytes,
-      incomingMetadata
+      bytes
     );
     const file = new StoredFile(md, this._persistence.getDiskPath(filePath));
     this._files.set(filePath, file);