+++ src/gcp/cloudfunctions.ts
@@ -465,6 +465,7 @@ export function endpointFromFunction(gcfFunction: CloudFunction): backend.Endpoi
   const [, project, , region, , id] = gcfFunction.name.split("/");
   let trigger: backend.Triggered;
   let uri: string | undefined;
+  let securityLevel: SecurityLevel | undefined;
   if (gcfFunction.labels?.["deployment-scheduled"]) {
     trigger = {
       scheduleTrigger: {},
@@ -476,6 +477,7 @@ export function endpointFromFunction(gcfFunction: CloudFunction): backend.Endpoi
   } else if (gcfFunction.httpsTrigger) {
     trigger = { httpsTrigger: {} };
     uri = gcfFunction.httpsTrigger.url;
+    securityLevel = gcfFunction.httpsTrigger.securityLevel;
   } else {
     trigger = {
       eventTrigger: {
@@ -504,6 +506,9 @@ export function endpointFromFunction(gcfFunction: CloudFunction): backend.Endpoi
   if (uri) {
     endpoint.uri = uri;
   }
+  if (securityLevel) {
+    endpoint.securityLevel = securityLevel;
+  }
   proto.copyIfPresent(
     endpoint,
     gcfFunction,
@@ -512,15 +517,21 @@ export function endpointFromFunction(gcfFunction: CloudFunction): backend.Endpoi
     "timeout",
     "minInstances",
     "maxInstances",
-    "vpcConnector",
-    "vpcConnectorEgressSettings",
     "ingressSettings",
     "labels",
     "environmentVariables",
     "secretEnvironmentVariables",
     "sourceUploadUrl"
   );
-
+  if (gcfFunction.vpcConnector) {
+    endpoint.vpc = { connector: gcfFunction.vpcConnector };
+    proto.renameIfPresent(
+      endpoint.vpc,
+      gcfFunction,
+      "egressSettings",
+      "vpcConnectorEgressSettings"
+    );
+  }
   return endpoint;
 }
 
@@ -575,6 +586,12 @@ export function functionFromEndpoint(
     gcfFunction.labels = { ...gcfFunction.labels, "deployment-taskqueue": "true" };
   } else {
     gcfFunction.httpsTrigger = {};
+    if (backend.isCallableTriggered(endpoint)) {
+      gcfFunction.labels = { ...gcfFunction.labels, "deployment-callabled": "true" };
+    }
+    if (endpoint.securityLevel) {
+      gcfFunction.httpsTrigger.securityLevel = endpoint.securityLevel;
+    }
   }
 
   proto.copyIfPresent(
@@ -585,12 +602,18 @@ export function functionFromEndpoint(
     "availableMemoryMb",
     "minInstances",
     "maxInstances",
-    "vpcConnector",
-    "vpcConnectorEgressSettings",
     "ingressSettings",
     "environmentVariables",
     "secretEnvironmentVariables"
   );
-
+  if (endpoint.vpc) {
+    proto.renameIfPresent(gcfFunction, endpoint.vpc, "vpcConnector", "connector");
+    proto.renameIfPresent(
+      gcfFunction,
+      endpoint.vpc,
+      "vpcConnectorEgressSettings",
+      "egressSettings"
+    );
+  }
   return gcfFunction;
 }