+++ src/test/deploy/functions/runtimes/discovery/v1alpha1.spec.ts
@@ -32,7 +32,7 @@ describe("backendFromV1Alpha1", () => {
       for (const [key, value] of Object.entries(invalidBackendTypes)) {
         it(`throws on invalid value for top-level key ${key}`, () => {
           const obj = {
-            requiredAPIs: {},
+            requiredAPIs: [],
             endpoints: {},
             [key]: value,
           };
@@ -311,7 +311,10 @@ describe("backendFromV1Alpha1", () => {
           },
         },
       };
-      const expected = backend.of({ ...DEFAULTED_ENDPOINT, scheduleTrigger });
+      const expected = backend.of({
+        ...DEFAULTED_ENDPOINT,
+        scheduleTrigger,
+      });
       const parsed = v1alpha1.backendFromV1Alpha1(yaml, PROJECT, REGION, RUNTIME);
       expect(parsed).to.deep.equal(expected);
     });
@@ -349,8 +352,10 @@ describe("backendFromV1Alpha1", () => {
         timeout: "60s",
         maxInstances: 20,
         minInstances: 1,
-        vpcConnector: "hello",
-        vpcConnectorEgressSettings: "ALL_TRAFFIC",
+        vpc: {
+          connector: "hello",
+          egressSettings: "ALL_TRAFFIC",
+        },
         ingressSettings: "ALLOW_INTERNAL_ONLY",
         serviceAccountEmail: "sa@",
       };