+++ src/test/gcp/cloudfunctionsv2.spec.ts
@@ -117,6 +117,10 @@ describe("cloudfunctionsv2", () => {
             },
           ],
         },
+        serviceConfig: {
+          ...CLOUD_FUNCTION_V2.serviceConfig,
+          environmentVariables: { FUNCTION_SIGNATURE_TYPE: "cloudevent" },
+        },
       };
       expect(
         cloudfunctionsv2.functionFromEndpoint(eventEndpoint, CLOUD_FUNCTION_V2_SOURCE)
@@ -144,8 +148,10 @@ describe("cloudfunctionsv2", () => {
         ...ENDPOINT,
         httpsTrigger: {},
         platform: "gcfv2",
-        vpcConnector: "connector",
-        vpcConnectorEgressSettings: "ALL_TRAFFIC",
+        vpc: {
+          connector: "connector",
+          egressSettings: "ALL_TRAFFIC",
+        },
         ingressSettings: "ALLOW_ALL",
         serviceAccountEmail: "inlined@google.com",
         labels: {
@@ -213,6 +219,7 @@ describe("cloudfunctionsv2", () => {
           minInstanceCount: 1,
           timeoutSeconds: 15,
           availableMemory: "128M",
+          environmentVariables: { FUNCTION_SIGNATURE_TYPE: "cloudevent" },
         },
       };
 
@@ -304,13 +311,15 @@ describe("cloudfunctionsv2", () => {
 
     it("should copy optional fields", () => {
       const extraFields: backend.ServiceConfiguration = {
-        vpcConnector: "connector",
-        vpcConnectorEgressSettings: "ALL_TRAFFIC",
         ingressSettings: "ALLOW_ALL",
         serviceAccountEmail: "inlined@google.com",
         environmentVariables: {
           FOO: "bar",
         },
+      };
+      const vpc = {
+        connector: "connector",
+        egressSettings: "ALL_TRAFFIC" as const,
       };
       expect(
         cloudfunctionsv2.endpointFromFunction({
@@ -318,6 +327,8 @@ describe("cloudfunctionsv2", () => {
           serviceConfig: {
             ...HAVE_CLOUD_FUNCTION_V2.serviceConfig,
             ...extraFields,
+            vpcConnector: vpc.connector,
+            vpcConnectorEgressSettings: vpc.egressSettings,
             availableMemory: "128M",
           },
           labels: {
@@ -330,6 +341,7 @@ describe("cloudfunctionsv2", () => {
         httpsTrigger: {},
         uri: RUN_URI,
         ...extraFields,
+        vpc,
         availableMemoryMb: 128,
         labels: {
           foo: "bar",