+++ CHANGELOG.md
@@ -4,4 +4,13 @@
 - Fixes broken functions:config:clone command (#4173).
 - Fixes issue where `auth:import` would fail when reading a JSON file (#4157).
 - Fixes issue where custom claims added in Auth Emulator UI was not properly shown (#4175).
+- Updates reserved environment variables for CF3 to include 'EVENTARC_CLOUD_EVENT_SOURCE' (#4196).
+- Fixes arg order for `firebase emulators:start --only storage` (#4195).
+- Fixes iOS auth for resumable uploads in Storage Emulator (#4184).
+- Fixes Storage Emulator crash on iOS auth error for resumable uploads (#4210).
+- Fixes bug where environment variable for gen 2 functions weren't updated on deploy (#4209).
+- Fixes an issue in the storage emulator where a file upload would trigger functions with a metadata update handler (#4213).
+- Fixes Storage Emulator rules resource evaluation (#4214).
+- Fixes bug where securityLevel is overwritten on https function re-deploys (#4208).
+- Fixes bug where functions emulator ignored functions.runtime option in firebase.json (#4207).
 - Updates functions init template to use latest versions of dependencies of functions init (#4177).